requirements.txt
discord.py
python-twitter
psycopg2